requirements.txt
web3==6.19.0
eth-account
click==8.1.7
py-solc-x==1.1.1
rich==13.7.1
python-dotenv==1.0.1
